Assignments/a6/final_project/requirements.txt
boto3==1.40.68
modal==1.2.1
numpy==2.3.5
pandas==2.3.3
psycopg2==2.9.11
pytest==9.0.1
PyYAML==6.0.3
sentence_transformers==5.1.2
torch==2.7.1+cu128
torch==2.10.0.dev20251020+cu128
tqdm==4.66.5
transformers==4.57.3
